move all the model shape drawing functions to general
